code cleanup, dependencies upgrade
