+++ package.json
@@ -38,7 +38,7 @@
     "typeorm": "^0.2.19",
     "typescript": "^3.6.4",
     "yargs": "^14.2.0",
-    "yn": "^2.0.0"
+    "yn": "^3.1.1"
   },
   "devDependencies": {
     "@types/chai": "^4.2.3",
@@ -57,9 +57,8 @@
     "@types/prettier": "^1.18.3",
     "@types/sinon": "^7.5.0",
     "@types/sqlite3": "^3.1.5",
-    "@types/yargs": "^12.0.1",
-    "@types/yn": "^3.1.0",
-    "@typescript-eslint/eslint-plugin": "^1.13.0",
+    "@types/yargs": "^13.0.3",
+    "@typescript-eslint/eslint-plugin": "^2.3.3",
     "@typescript-eslint/parser": "^2.3.3",
     "@typescript-eslint/typescript-estree": "^2.3.3",
     "chai": "^4.2.0",
