+++ test/utils/GeneralTestUtils.ts
@@ -41,7 +41,7 @@ export async function createMSSQLModels(
         password: String(process.env.MSSQL_Password),
         databaseType: "mssql",
         schemaName: "dbo,sch1,sch2",
-        ssl: yn(process.env.MSSQL_SSL)
+        ssl: yn(process.env.MSSQL_SSL, { default: false })
     };
     await driver.ConnectToServer(connectionOptions);
     connectionOptions.databaseName = String(process.env.MSSQL_Database);
@@ -92,7 +92,7 @@ export async function createPostgresModels(
         password: String(process.env.POSTGRES_Password),
         databaseType: "postgres",
         schemaName: "public,sch1,sch2",
-        ssl: yn(process.env.POSTGRES_SSL)
+        ssl: yn(process.env.POSTGRES_SSL, { default: false })
     };
     await driver.ConnectToServer(connectionOptions);
     connectionOptions.databaseName = String(process.env.POSTGRES_Database);
@@ -176,7 +176,7 @@ export async function createMysqlModels(
         password: String(process.env.MYSQL_Password),
         databaseType: "mysql",
         schemaName: "ignored",
-        ssl: yn(process.env.MYSQL_SSL)
+        ssl: yn(process.env.MYSQL_SSL, { default: false })
     };
     await driver.ConnectToServer(connectionOptions);
 
@@ -218,7 +218,7 @@ export async function createMariaDBModels(
         password: String(process.env.MARIADB_Password),
         databaseType: "mariadb",
         schemaName: "ignored",
-        ssl: yn(process.env.MARIADB_SSL)
+        ssl: yn(process.env.MARIADB_SSL, { default: false })
     };
     await driver.ConnectToServer(connectionOptions);
 
@@ -262,7 +262,7 @@ export async function createOracleDBModels(
         password: String(process.env.ORACLE_PasswordSys),
         databaseType: "oracle",
         schemaName: String(process.env.ORACLE_Username),
-        ssl: yn(process.env.ORACLE_SSL)
+        ssl: yn(process.env.ORACLE_SSL, { default: false })
     };
     await driver.ConnectToServer(connectionOptions);
     connectionOptions.user = String(process.env.ORACLE_Username);
